FIR-2777 Use present? to check for config values
Before, the system would just evaluate the truthiness of the config. Unfortunately, a blank string ("") is truthy, so a blank value for user agent name or rets version would result in those headers getting added to the RETS request. With those headers present, but no values in them, the RETS servers would get confused and demand that we include a User-Agent header even though they don't actually require the User-Agent header. Sending a request with no User-Agent header returns successfully, but by virtue of having the User-Agent key present in the headers, we confuse the RETS servers. This commit uses present? to check for values, which will return false if a blank string is present or the value is nil.